pyproject.toml
[tool.poetry]
name = "fym-neupic-queue"
version = "0.1.0"
description = ""
authors = ["Alex Malykh <alexeyheather@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12.4"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
